+++ extension-chromium/apply-theme-early.js
@@ -1,0 +1,1 @@
+../extension/apply-theme-early.js