+++ extension-chromium/blocked.js
@@ -1,0 +1,1 @@
+../extension/blocked.js