+++ extension-chromium/pages-dev.js
@@ -1,0 +1,1 @@
+../extension/pages-dev.js